refactor(frontend): separate UI for selecting ingredients into a distinct component
