+++ apps/frontend/src/routes/_app/upload.tsx
@@ -1,7 +1,7 @@
-import { IconCircleCheck, IconCircleX, IconPlus } from '@tabler/icons-react'
 import { createFileRoute } from '@tanstack/react-router'
 import { type ChangeEvent, useMemo, useState } from 'react'
 import { LinkButton } from '../../components/common/LinkButton'
+import { FoodSelect } from '../../components/upload/FoodSelect'
 import { ImagePicker } from '../../components/upload/ImagePicker'
 import { apiClient } from '../../lib/apiClient'
 import type { FoodImage } from '../../types/FoodTypes'
@@ -41,15 +41,6 @@ const Upload = () => {
     setFoodImages((prev) => prev.filter((_, i) => i !== index))
   }
 
-  const toggleFoodSelect = (food: string) => {
-    setSelectedFoods((prev) => {
-      if (prev.includes(food)) {
-        return prev.filter((f) => f !== food)
-      }
-      return [...prev, food]
-    })
-  }
-
   const foods = useMemo(
     () => dropDuplicates(foodImages.flatMap((image) => image.foods)),
     [foodImages],
@@ -62,35 +53,7 @@ const Upload = () => {
         handleRemove={handleFileRemove}
         foodImages={foodImages}
       />
-      <div className="flex flex-wrap gap-x-1 gap-y-2">
-        {foods.map((food) => (
-          <div
-            key={food}
-            className="flex cursor-pointer items-center gap-x-2 rounded-full border py-1 pr-4 pl-2"
-            style={{
-              borderColor: selectedFoods.includes(food) ? '#6c8' : 'gray',
-              backgroundColor: selectedFoods.includes(food) ? '#6c82' : 'transparent',
-            }}
-            onClick={() => toggleFoodSelect(food)}
-            onKeyDown={() => toggleFoodSelect(food)}
-          >
-            {selectedFoods.includes(food) ? (
-              <IconCircleCheck size={20} color="#6c8" className="aspect-square" />
-            ) : (
-              <IconCircleX size={20} color="gray" className="aspect-square" />
-            )}
-            <span>{food}</span>
-          </div>
-        ))}
-        {foods.length !== 0 && (
-          <button
-            className="box-content flex h-[1lh] w-[1lh] items-center justify-center rounded-full border border-[#6c8] bg-[#6c8] px-1 py-1"
-            type="button"
-          >
-            <IconPlus size={20} color="#fff" className="aspect-square" />
-          </button>
-        )}
-      </div>
+      <FoodSelect foods={foods} selectedFoods={selectedFoods} setSelectedFoods={setSelectedFoods} />
       {foods.length !== 0 && (
         <LinkButton to="/recipe" search={{ foods: selectedFoods }}>
           レシピを検索
